Rename the data pointer variable from data to pData.  Also fixed a buffer size allocation issue where the wrong size was used to create the buffer.
In lines 15 and 36, the buffer size was determined by referencing the vertices and indices as
sizeof(pBodyManager->getPSquare()->getPVertices()[0]) and
sizeof(pBodyManager0>getPSquare()->getPIndices()[0]), however, because these are pointers to vectors, the correct way to retrieve the sizes is through dereferencing.  ie.
sizeof((*pBodyManager->getPSquare()->getPVertices())[0]) and
sizeof((*pBodyManager0>getPSquare()->getPIndices())[0]).

This has now been fixed.